Refactor and correct data seeding logic
- Centralized all seeding operations into IdentityDataSeeder.
- Ensured driver user accounts are created before trucks to allow for correct association.
- Modified destination region seeding to assign all regions to the Surrey branch.
- Deleted redundant seeding file and updated Program.cs.
