+++ Program.cs
@@ -160,7 +160,6 @@ using (var scope = app.Services.CreateScope())
     var dbContext = scope.ServiceProvider.GetRequiredService<ApplicationDbContext>();
     var roleManager = scope.ServiceProvider.GetRequiredService<RoleManager<ApplicationRole>>();
     var userManager = scope.ServiceProvider.GetRequiredService<UserManager<ApplicationUser>>();
-    await CMetalsWS.Data.Seed.SeedDestinationRegions.RunAsync(dbContext, roleManager, userManager);
 }
 
 // Pipeline
